% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSRDLdb.R
\name{dep.SRDLdb}
\alias{dep.SRDLdb}
\title{\code{dep.SRDLdb} retrieves the entire deployments table from a SMRU Access database}
\usage{
dep.SRDLdb(theDB = "ct34", thePath = "C:/MamVisAD/SMRUAccessDatabases")
}
\arguments{
\item{theDB}{Name of database as defined in ODBC data sources}

\item{thePath}{Path to folder where database is located, if running in 64-bit R.}
}
\value{
Returns a data frame with all deployment metadata
}
\description{
\code{dep.SRDLdb} retrieves the entire deployments table from a SMRU Access database
}
\details{
The database must first be set up as a DSN data source.
  Note! On Windows, by default only 32-bit data sources can be used
  and R (or RStudio) has to be run in 32-bit mode for these functions to work.
  However, functionality has now been extended (experimental for now) to run under 64-bit R.
  For this to worl, the path to the database must be specified.
}
\seealso{
\code{\link{tables.SRDLdb}} for database table names,
  \code{\link{fields.SRDLdb}} for table field names,
  \code{\link{ref.SRDLdb}} for ref ID codes,
  \code{\link{get.SRDLdb}} for querying and retrieving data from database table
}
\author{
Martin Biuw
}
