% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSRDLdb.R
\name{fields.SRDLdb}
\alias{fields.SRDLdb}
\title{\code{fields.SRDLdb} lists field names form specified table in a SMRU Access database}
\usage{
fields.SRDLdb(
  theDB = "ct34",
  theTable = "dive",
  thePath = "C:/MamVisAD/SMRUAccessDatabases"
)
}
\arguments{
\item{theDB}{Name of database as defined in ODBC data sources}

\item{theTable}{Name of database table}

\item{thePath}{Path to folder where database is located, if running in 64-bit R.}
}
\value{
Returns a character vector with field names from specified table
}
\description{
\code{fields.SRDLdb} lists field names form specified table in a SMRU Access database
}
\details{
The database must first be set up as a DSN data source.
  Note! On Windows, by default only 32-bit data sources can be used
  and R (or RStudio) has to be run in 32-bit mode for these functions to work.
  However, functionality has now been extended (experimental for now) to run under 64-bit R.
  For this to worl, the path to the database must be specified.
}
\seealso{
\code{\link{tables.SRDLdb}} for database table names,
  \code{\link{ref.SRDLdb}} for ref ID codes,
  \code{\link{dep.SRDLdb}} for retrieving deployments metadata,
  \code{\link{get.SRDLdb}} for querying and retrieving data from database table

Other SMRU SRDL database functions: 
\code{\link{fields.SRDLpg}()},
\code{\link{ref.SRDLdb}()},
\code{\link{ref.SRDLpg}()},
\code{\link{tables.SRDLdb}()},
\code{\link{tables.SRDLpg}()}
}
\author{
Martin Biuw
}
\concept{SMRU SRDL database functions}
