% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSRDLdb.R
\name{get.all.SRDLdb}
\alias{get.all.SRDLdb}
\title{\code{get.all.SRDLdb} retrieves entire SMRU Access database}
\usage{
get.all.SRDLdb(theDB = "ct34", thePath = "C:/MamVisAD/SMRUAccessDatabases")
}
\arguments{
\item{theDB}{Name of database as defined in ODBC data sources}

\item{thePath}{Path to folder where database is located, if running in 64-bit R.}
}
\value{
Returns a list of data frames, each corresponding to a SMRU Access data table.
}
\description{
\code{get.all.SRDLdb} retrieves entire SMRU Access database
}
\details{
The database must first be set up as a DSN data source.
  Note! On Windows, by default only 32-bit data sources can be used
  and R (or RStudio) has to be run in 32-bit mode for these functions to work.
  However, functionality has now been extended (experimental for now) to run under 64-bit R.
  For this to worl, the path to the database must be specified.
}
\examples{
dive <- get.all.SRDLdb('ct34')
}
\seealso{
\code{\link{tables.SRDLdb}} for database table names,
  \code{\link{fields.SRDLdb}} for table field names,
  \code{\link{ref.SRDLdb}} for ref ID codes,
  \code{\link{dep.SRDLdb}} for retrieving deployments metadata
  \code{\link{get.SRDLdb}} for retrieving single tables based on given selection criteria
}
\author{
Martin Biuw
}
