% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSRDLpg.R
\name{get.all.SRDLpg}
\alias{get.all.SRDLpg}
\title{\code{get.all.SRDLpg} retrieves entire SMRU PostgreSQL database}
\usage{
get.all.SRDLpg(theDB = "SRDL", theDep = "ct34")
}
\arguments{
\item{theDB}{Name of database}
}
\value{
Returns a list of data frames, each corresponding to a SMRU Access data table.
}
\description{
\code{get.all.SRDLpg} retrieves entire SMRU PostgreSQL database
}
\details{
DO not run this function without a specific deployment specified,
unless your PostgreSQL database is of reasonable size. Otherwise it will
take a very long time to load, or potentially crash.
}
\examples{
dive <- get.all.SRDLpg('SRDL')
}
\seealso{
\code{\link{tables.SRDLpg}} for database table names,
  \code{\link{fields.SRDLpg}} for table field names,
  \code{\link{ref.SRDLpg}} for ref ID codes,
  \code{\link{dep.SRDLpg}} for retrieving deployments metadata
  \code{\link{get.SRDLpg}} for retrieving single tables based on given selection criteria
}
\author{
Martin Biuw
}
