% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSRDLdb.R
\name{get.SRDLdb}
\alias{get.SRDLdb}
\title{\code{get.SRDLdb} runs a query to fetch data from SMRU Access database table, based on set criteria}
\usage{
get.SRDLdb(
  theDB = "ct34",
  theTable = "dive",
  theFields = "All",
  theRef = "ct34-2447-08",
  thePath = "C:/MamVisAD/SMRUAccessDatabases"
)
}
\arguments{
\item{theDB}{Name of database as defined in ODBC data sources}

\item{theTable}{Name of database table}

\item{theFields}{Vector of names or numbers representing which fields to be fetched,
or "All" to get all tables.}

\item{theRef}{ref code of specific animal for which data should be fetched.
Can be either "All" for entire table, or one specific ref code.
There is currently no support for subsets of multiple ref codes}

\item{thePath}{Path to folder where database is located, if running in 64-bit R.}
}
\value{
Returns a data frame with data from table following specified selection criteria.
}
\description{
\code{get.SRDLdb} runs a query to fetch data from SMRU Access database table, based on set criteria
}
\details{
The database must first be set up as a DSN data source.
  Note! On Windows, by default only 32-bit data sources can be used
  and R (or RStudio) has to be run in 32-bit mode for these functions to work.
  However, functionality has now been extended (experimental for now) to run under 64-bit R.
  For this to worl, the path to the database must be specified.
}
\examples{
dive <- get.SRDLdb('ct34', 'dive', c(c(1:35), c(79:80)), 'ct34-2447-08')
dive <- get.SRDLdb('ct34', 'dive', c('ref', 'DE_DATE', 'SURF_DUR', 'DIVE_DUR', 'MAX_DEP')
}
\seealso{
\code{\link{tables.SRDLdb}} for database table names,
  \code{\link{fields.SRDLdb}} for table field names,
  \code{\link{ref.SRDLdb}} for ref ID codes,
  \code{\link{dep.SRDLdb}} for retrieving deployments metadata
}
\author{
Martin Biuw
}
