% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSRDLpg.R
\name{get.SRDLpg}
\alias{get.SRDLpg}
\title{\code{get.SRDLpg} runs a query to fetch data from SMRU PostgreSQL database table, based on set criteria}
\usage{
get.SRDLpg(
  theDB = "SRDL",
  theTable = "dive",
  theFields = "All",
  theDep = "ct34",
  theRef = "All",
  theHost = "localhost",
  thePort = 5432,
  theUser = "postgres",
  thePwd = "12345"
)
}
\arguments{
\item{theDB}{Name of database}

\item{theTable}{Name of database table}

\item{theFields}{Vector of names or numbers representing which fields to be fetched,
or "All" to get all tables.}

\item{theDep}{Either an exact name of a deployment program to be loaded,
or a wildcard for similarly named deployment programs.}

\item{theRef}{ref code of specific animal for which data should be fetched.
Can be either "All" for entire table, or one specific ref code.
There is currently no support for subsets of multiple ref codes}
}
\value{
Returns a data frame with data from table following specified selection criteria.
}
\description{
\code{get.SRDLpg} runs a query to fetch data from SMRU PostgreSQL database table, based on set criteria
}
\examples{
dive <- get.SRDLpg('SRDL', 'dive', 'All', 'ct34-2447-08')
dive <- get.SRDLpg('SRDL', 'dive', c('ref', 'de_date', 'surf_dur', 'dive_dur', 'max_dep')
}
\seealso{
\code{\link{tables.SRDLpg}} for database table names,
  \code{\link{fields.SRDLpg}} for table field names,
  \code{\link{ref.SRDLpg}} for ref ID codes,
  \code{\link{dep.SRDLpg}} for retrieving deployments metadata
}
\author{
Martin Biuw
}
