% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSRDLdb.R
\name{mdb2pg}
\alias{mdb2pg}
\title{\code{mdb2pg} Write entire SMRU Access database to PostgreSQL}
\usage{
mdb2pg(mdb = "ct34", pg = "SRDL", append = F)
}
\arguments{
\item{mdb}{Either a character string giving the name of MS Access database
as defined in ODBC data sources or the name of an R list which is the result of
reading an entire SMRU Access deployments database (using the \code{\link{get.all.SRDLdb}} function)}

\item{pg}{Name of PostgreSQL database into which tables should be written}

\item{append}{Logical, indicating whether to create table (default) or append to existing table}
}
\value{
Returns nothing, but re-creates a PostgreSQL representation of the MS Access database,
potentially merging several deployments into single set of tables.
}
\description{
\code{mdb2pg} Write entire SMRU Access database to PostgreSQL
}
\details{
The source MS Access database must first be set up as a DSN data source.
  Note! On Windows, only 32-bit data sources can be used,
  and R (or RStudio) has to be run in 32-bit mode for these functions to work.
  In addition, a PostgreSQL/PostGIS database must exist that can hold the tables.
}
\examples{
mdb2pg('ct34')
for(i in all.mdb) {
  cat('Converting database', i, '(', match(i, all.mdb), 'of', length(all.mdb), ')\n')
  flush.console()
  if(match(i, all.db)==1) {
    mdb2pg(i)
  } else{
    mdb2pg(i)
  }
}
}
\seealso{
\code{\link{tables.SRDLdb}} for database table names,
  \code{\link{fields.SRDLdb}} for table field names,
  \code{\link{ref.SRDLdb}} for ref ID codes,
  \code{\link{dep.SRDLdb}} for retrieving deployments metadata
  \code{\link{get.SRDLdb}} for retrieving single tables based on given selection criteria
  \code{\link{get.all.SRDLdb}} for retrieving entire SMRU Access database
}
\author{
Martin Biuw
}
