% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qcDive.R
\name{qcDive}
\alias{qcDive}
\title{\code{qcDive} Checks consistency of dive records}
\usage{
qcDive(d.dat = dive, flag.only = T, max.v.speed = 5)
}
\arguments{
\item{d.dat}{Dive data to be quality controlled}

\item{flag.only}{logical specifying if erroneous records should be flagged or deleted}

\item{max.v.speed}{Threshold for maximum allowable rate of change in depth
between inflection points (in m/s)}
}
\value{
Returns a data frame with either an added column with qc flag
(0 for OK, 1 for error), or with erroneous records deleted
  or an interactive dygraph (if \code{type} is set to 'dygraph')
}
\description{
\code{qcDive} Checks consistency of dive records
}
\details{
the function checks for consistency in time sequences between
inflections points and unrealistic depth changes between inflection points.
}
\examples{
dive.deleted <- qcDive(dive)
dive.flagged <- qcDive(flagged.only=T)
}
\seealso{
\code{\link{fields.SRDLdb}} for table field names,
  \code{\link{ref.SRDLdb}} for ref ID codes,
  \code{\link{dep.SRDLdb}} for retrieving deployments metadata,
  \code{\link{get.SRDLdb}} for querying and retrieving data from database table
  \code{\link{strip.SRDL}}makes a stripchart of SRDL dive data

Other SMRU SRDL functions: 
\code{\link{map.SRDL}()},
\code{\link{strip.SRDL}()}
}
\author{
Martin Biuw
}
\concept{SMRU SRDL functions}
