% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSRDLpg.R
\name{ref.SRDLpg}
\alias{ref.SRDLpg}
\title{\code{ref.SRDLpg} lists all ref codes (animal ID) in a SMRU PostgreSQL database}
\usage{
ref.SRDLpg(
  theDB = "SRDL",
  theHost = "localhost",
  thePort = 5432,
  theUser = "postgres",
  thePwd = "12345"
)
}
\arguments{
\item{theDB}{Name of database}
}
\value{
Returns a character vector with all ref codes
}
\description{
\code{ref.SRDLpg} lists all ref codes (animal ID) in a SMRU PostgreSQL database
}
\seealso{
\code{\link{tables.SRDLpg}} for database table names,
  \code{\link{fields.SRDLpg}} for table field names,
  \code{\link{dep.SRDLpg}} for retrieving deployments metadata,
  \code{\link{get.SRDLpg}} for querying and retrieving data from database table

Other SMRU SRDL database functions: 
\code{\link{fields.SRDLdb}()},
\code{\link{fields.SRDLpg}()},
\code{\link{ref.SRDLdb}()},
\code{\link{tables.SRDLdb}()},
\code{\link{tables.SRDLpg}()}
}
\author{
Martin Biuw
}
\concept{SMRU SRDL database functions}
