% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripSRDL.R
\name{strip.SRDL}
\alias{strip.SRDL}
\title{\code{strip.SRDL} makes a stripchart of SRDL dive data}
\usage{
strip.SRDL(
  dat = hp4,
  ctd = F,
  d.sum = T,
  theRef = levels(hp4$dive$REF)[1],
  type = "normal",
  tz = "UTC",
  QQ = F
)
}
\arguments{
\item{QQ}{Logical, should dive quality check be performed. Note! Only works for 4-point dive data!}

\item{dive}{Logical, should detailed dive data be plotted?}

\item{c.dat}{Logical, should max depth from CTD data be plotted?}

\item{s.dat}{Logical, should dive summary data be plotted?}
}
\value{
Returns a stripchart plot, either as standard R plot
  or an interactive dygraph (if \code{type} is set to 'dygraph')
}
\description{
\code{strip.SRDL} makes a stripchart of SRDL dive data
}
\details{
The dygraphs option depends on the \code{\link{dygraph}} and \code{\link{xts}}
  functions, and the relevant packages (\code{dygraphs} & \code{xts}) must be installed for the dygraphs option to work.
}
\examples{
strip.SRDLdb(dive, ctd, summ)
strip.SRDLdb(dive, ctd, summ, type='dygraph')
#' @importFrom xts xts
}
\seealso{
\code{\link{fields.SRDLdb}} for table field names,
  \code{\link{ref.SRDLdb}} for ref ID codes,
  \code{\link{dep.SRDLdb}} for retrieving deployments metadata,
  \code{\link{get.SRDLdb}} for querying and retrieving data from database table

Other SMRU SRDL functions: 
\code{\link{map.SRDL}()},
\code{\link{qcDive}()}
}
\author{
Martin Biuw
}
\concept{SMRU SRDL functions}
