% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSRDLpg.R
\name{tables.SRDLpg}
\alias{tables.SRDLpg}
\title{\code{tables.SRDLpg} lists the tables in a SMRU PostgreSQL database}
\usage{
tables.SRDLpg(
  theDB = "SRDL",
  theHost = "localhost",
  thePort = 5432,
  theUser = "postgres",
  thePwd = "12345"
)
}
\arguments{
\item{theDB}{Name of (or path to) the pg database}
}
\value{
Returns a character vector with database table names
}
\description{
\code{tables.SRDLpg} lists the tables in a SMRU PostgreSQL database
}
\seealso{
\code{\link{fields.SRDLpg}} for table field names,
  \code{\link{ref.SRDLpg}} for ref ID codes,
  \code{\link{dep.SRDLpg}} for retrieving deployments metadata,
  \code{\link{get.SRDLpg}} for querying and retrieving data from database table

Other SMRU SRDL database functions: 
\code{\link{fields.SRDLdb}()},
\code{\link{fields.SRDLpg}()},
\code{\link{ref.SRDLdb}()},
\code{\link{ref.SRDLpg}()},
\code{\link{tables.SRDLdb}()}
}
\author{
Martin Biuw
}
\concept{SMRU SRDL database functions}
